move in convert_model.coreml
